+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-1.2.4...master
 
+### Fixed
+- No circuit breaker by default.
+
 ## [1.2.4] - 2021-04-15
 [1.2.4]: https://github.com/atlassian-labs/db-replica/compare/release-1.2.2...release-1.2.4